requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
asttokens==3.0.0
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
gdown==5.2.0
idna==3.11
ipykernel==7.0.1
ipython==9.6.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
jinja2==3.1.6
jupyter-client==8.6.3
jupyter-core==5.9.1
jupyterlab-widgets==3.0.15
kagglehub==0.3.13
kiwisolver==1.4.9
markupsafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==12.0.0
platformdirs==4.5.0
prompt-toolkit==3.0.52
psutil==7.1.0
pure-eval==0.2.3
pydantic==2.12.3
pydantic-core==2.41.4
pygments==2.19.2
pyparsing==3.2.5
pysocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
requests==2.32.5
scipy==1.16.2
seaborn==0.13.2
setuptools==80.9.0
simsimd==6.5.3
six==1.17.0
soupsieve==2.8
stack-data==0.6.3
stringzilla==4.2.1
sympy==1.14.0
torch==2.9.0
torchsummary==1.5.1
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
widgetsnbextension==4.0.14
